feat: Add LFU cache implementation
This commit adds a new file `100-lfu_cache.py` that contains the implementation of the `LFUCache` class. This class extends the `BaseCaching` class and provides methods for adding items to the cache (`put`) and retrieving items from the cache (`get`). The LFU (Least Frequently Used) caching algorithm is used to manage the cache data. This addition enhances the caching functionality in the project.