feat: Add i18n support to Flask app
This commit adds the necessary code changes to enable internationalization (i18n) support in the Flask app. It includes the addition of the Babel library, configuration of supported languages, and rendering of a basic template. This enhancement allows for the localization of the app's content based on the user's preferred language.